+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - none
 
 # v2022.10.1-beta
+- chart - use different color for category - short
 - sqlmath - re-enable sql-function squared()
 
 # v2022.9.20